# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rental Dashboard API",
    "description": "RESTful API for managing car rental operations including cars, bookings, customers, expenses, maintenance, and incidents.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3010",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token or API key authentication"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["admin", "staff", "readonly"] },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "Settings": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "agency_name": { "type": "string" },
          "currency": { "type": "string" },
          "vat_percent": { "type": "number", "format": "float" },
          "logo_path": { "type": "string", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "LateFeeRule": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "hours_from": { "type": "integer" },
          "hours_to": { "type": "integer", "nullable": true },
          "fee_amount": { "type": "number", "format": "float" },
          "fee_type": { "type": "string", "enum": ["flat", "percentage"] },
          "is_active": { "type": "integer" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/settings": {
      "get": {
        "summary": "Get agency settings",
        "tags": ["Settings"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Settings",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Settings" }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update agency settings",
        "tags": ["Settings"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "agency_name": { "type": "string" },
                  "currency": { "type": "string" },
                  "vat_percent": { "type": "number", "minimum": 0, "maximum": 100 },
                  "logo_path": { "type": "string", "nullable": true }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Settings updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Settings" }
              }
            }
          },
          "400": { "description": "Invalid input" },
          "403": { "description": "Forbidden - Admin only" }
        }
      }
    },
    "/api/settings/late-fee-rules": {
      "get": {
        "summary": "Get late fee rules",
        "tags": ["Settings"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of late fee rules",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/LateFeeRule" } }
              }
            }
          },
          "403": { "description": "Forbidden - Admin only" }
        }
      },
      "post": {
        "summary": "Create late fee rule",
        "tags": ["Settings"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["hours_from", "fee_amount", "fee_type"],
                "properties": {
                  "hours_from": { "type": "integer", "minimum": 0 },
                  "hours_to": { "type": "integer", "nullable": true },
                  "fee_amount": { "type": "number", "minimum": 0 },
                  "fee_type": { "type": "string", "enum": ["flat", "percentage"] },
                  "is_active": { "type": "boolean", "default": true }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Late fee rule created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LateFeeRule" }
              }
            }
          },
          "400": { "description": "Invalid input" },
          "403": { "description": "Forbidden - Admin only" }
        }
      }
    },
    "/api/settings/late-fee-rules/{id}": {
      "put": {
        "summary": "Update late fee rule",
        "tags": ["Settings"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "hours_from": { "type": "integer" },
                  "hours_to": { "type": "integer", "nullable": true },
                  "fee_amount": { "type": "number" },
                  "fee_type": { "type": "string", "enum": ["flat", "percentage"] },
                  "is_active": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Late fee rule updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LateFeeRule" }
              }
            }
          },
          "400": { "description": "Invalid input" },
          "403": { "description": "Forbidden - Admin only" },
          "404": { "description": "Late fee rule not found" }
        }
      },
      "delete": {
        "summary": "Delete late fee rule",
        "tags": ["Settings"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Late fee rule deleted",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "ok": { "type": "boolean" } } }
              }
            }
          },
          "403": { "description": "Forbidden - Admin only" },
          "404": { "description": "Late fee rule not found" }
        }
      }
    },
...
    "/api/users": {
      "get": {
        "summary": "List all users",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/User" } }
              }
            }
          },
          "403": { "description": "Forbidden - Admin only" }
        }
      },
      "post": {
        "summary": "Create a new user",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "role"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 6 },
                  "role": { "type": "string", "enum": ["admin", "staff", "readonly"] }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "400": { "description": "Invalid input" },
          "403": { "description": "Forbidden - Admin only" }
        }
      }
    },
...
  },
...
}